#3: Refactored names to match standard template
