+++ lib/catalog/widgets/playground.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_quest/property_fields/boolean_field.dart';
+import 'package:flutter_quest/property_fields/list_field.dart';
 import 'package:flutter_quest/property_fields/number_field.dart';
 import 'package:flutter_quest/widgets/core/propery_builder.dart';
 
@@ -14,18 +15,34 @@ class PropertyPlayground extends StatelessWidget {
         final height = provider.heightField();
         final width = provider.widthField();
         final isRed = provider.booleanField(id: "isRed", title: "Is Red");
+        final text = provider.listField<String>(
+          title: "String",
+          id: 'string',
+          values: [
+            "One",
+            "Two",
+            "Three",
+            "Four",
+          ],
+        );
 
         final color = isRed == null
             ? Colors.green
             : isRed
                 ? Colors.red
                 : Colors.blue;
+
         return Container(
           height: height,
           width: width,
           decoration: BoxDecoration(
             color: color,
           ),
+          child: text == null
+              ? null
+              : Center(
+                  child: Text(text),
+                ),
         );
       },
     );
